ec2spots_workshop/requirements.txt
aws-cdk-lib==2.95.0
constructs>=10.0.0,<11.0.0
boto3==1.28.83
requests==2.31.0